+++ etcd/v3/Action.hpp
@@ -30,8 +30,8 @@ namespace etcdv3
   {
     ActionParameters();
     bool withPrefix;
-    int revision;
-    int old_revision;
+    int64_t revision;
+    int64_t old_revision;
     int64_t lease_id = 0;  // no lease
     int ttl;
     int limit;
